Move event handling to separate class
