+++ Source code/MscModApi/MscModApi.csproj
@@ -87,6 +87,8 @@
     <Compile Include="PaintingSystem\SerializableColor.cs" />
     <Compile Include="Parts\BasicPart.cs" />
     <Compile Include="Parts\DerivablePart.cs" />
+    <Compile Include="Parts\EventSupportingBasicPart.cs" />
+    <Compile Include="Parts\ReplacePart\GamePart.cs" />
     <Compile Include="Parts\ReplacePart\NewPart.cs" />
     <Compile Include="Parts\ReplacePart\OldPart.cs" />
     <Compile Include="Parts\ReplacePart\ReplacementPart.cs" />
